@@ -1,15 +1,19 @@
 name: vy_code_generator
 description: A package for dart code generator. A simpler alternative to code_generator.
-version: 0.1.0
+version: 0.2.0
 # homepage: https://www.example.com
+publish_to: none
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  vy_string_utils: ^0.3.2
-  dart_style: ^1.3.3
+  vy_string_utils: ^0.4.0-nullsafety
+  dart_style: #^2.0.0
+    git: 
+      url: "https://github.com/dart-lang/dart_style.git"
+      ref: a1d65dd9c4b83c9702ba409465879b94caeb04ea
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.0
+  pedantic: ^1.11.0
+  test: ^1.16.4
