@@ -1,7 +1,7 @@
 import '../identifier.dart';
 
 mixin Identified {
-  Identifier id;
+  Identifier? id;
 
   bool get isIdentified => id != null;
 }
