@@ -14,7 +14,7 @@ import 'utils/string_buffer_extension.dart';
 
 class DartClass extends NamedElement with DartType, Annotated {
   bool isAbstract = false;
-  DartClass extend;
+  DartClass? extend;
   final List<NamedElement> _withList = <NamedElement>[];
   final List<DartClass> _implementList = <DartClass>[];
 
@@ -27,47 +27,57 @@ class DartClass extends NamedElement with DartType, Annotated {
     type = id.id;
   }
   DartClass(String className) : super(Identifier(className)) {
-    type = id.id;
+    type = id?.id;
   }
   DartClass.fromTextualContent(String text) : super.fromTextualContent(text);
 
-  List<Constructor> get constructors => _constructors;
+  List<Constructor>? get constructors => _constructors;
   List<DartClass> get implementingClasses => _implementList;
 
   void addMixin(NamedElement mixin) {
-    var existingMixin = _withList.firstWhere(
-        (element) => element.id.id == mixin.id.id,
-        orElse: () => null);
-    if (existingMixin == null) {
+    if (mixin.id == null) {
+      throw ArgumentError('Cannot add mixin with no id');
+    }
+    try {
+      _withList.firstWhere((element) => mixin.id!.id == element.id?.id);
+    } on StateError {
+      // If not found
       _withList.add(mixin);
       return;
     }
-    throw StateError('The mixin "${mixin.id.id}" is already present '
-        'in class "${id.id}"');
+    throw StateError('The mixin "${mixin.id!.id}" is already present '
+        'in class "${id?.id == null ? '' : id!.id}"');
   }
 
   void addImplements(DartClass implementClass) {
-    var existingImplements = _implementList.firstWhere(
-        (element) => element.id.id == implementClass.id.id,
-        orElse: () => null);
-    if (existingImplements == null) {
+    if (implementClass.id == null) {
+      throw ArgumentError('Cannot add implements element with no id');
+    }
+    try {
+      _implementList
+          .firstWhere((element) => implementClass.id!.id == element.id?.id);
+    } on StateError {
+      // If not found
       _implementList.add(implementClass);
       return;
     }
-    throw StateError('The class "${implementClass.id.id}" is already present '
-        'in class "${id.id}"');
+    throw StateError('The class "${implementClass.id!.id}" is already present '
+        'in class "${id?.id == null ? '' : id!.id}"');
   }
 
   void addFieldSeparator() => addField(Variable.fromTextualContent('\n'));
 
   void addField(Variable field) {
-    var existingField = _fields.firstWhere(
-        (element) =>
-            filled(element.id?.id) &&
-            filled(field.id?.id) &&
-            element.id.id == field.id.id,
-        orElse: () => null);
-    if (existingField == null) {
+    if (field.id == null) {
+      throw ArgumentError('Cannot add field with no id');
+    }
+    try {
+      _fields.firstWhere((element) =>
+          filled(element.id?.id) &&
+          filled(field.id?.id) &&
+          element.id!.id == field.id!.id);
+    } on StateError {
+      // If not found
       field.parent = this;
       if (library != null) {
         field.library = library;
@@ -75,15 +85,18 @@ class DartClass extends NamedElement with DartType, Annotated {
       _fields.add(field);
       return;
     }
-    throw StateError('The field "${field.id.id}" is already present '
-        'in class "${id.id}"');
+    throw StateError('The field "${field.id!.id}" is already present '
+        'in class "${id?.id == null ? '' : id!.id}"');
   }
 
   void addConstructor(Constructor constructor) {
-    var existingConstructor = _constructors.firstWhere(
-        (element) => element.named == constructor.named,
-        orElse: () => null);
-    if (existingConstructor == null) {
+    if (constructor.named == null) {
+      throw ArgumentError('Cannot add constructor with no name');
+    }
+    try {
+      _constructors.firstWhere((element) => element.named == constructor.named);
+    } on StateError {
+      // If not found
       constructor.parent = this;
       if (library != null) {
         constructor.library = library;
@@ -93,14 +106,16 @@ class DartClass extends NamedElement with DartType, Annotated {
     }
     throw StateError(
         'The constructor "${constructor.named}" is already present '
-        'in class "${id.id}"');
+        'in class "${id?.id == null ? '' : id!.id}"');
   }
 
   void addFactory(DartFactory factory) {
-    var existingFactory = _factories.firstWhere(
-        (element) => element.named == factory.named,
-        orElse: () => null);
-    if (existingFactory == null) {
+    if (factory.named == null) {
+      throw ArgumentError('Cannot add factory with no name');
+    }
+    try {
+      _factories.firstWhere((element) => element.named == factory.named);
+    } on StateError {
       factory.parent = this;
       if (library != null) {
         factory.library = library;
@@ -109,19 +124,21 @@ class DartClass extends NamedElement with DartType, Annotated {
       return;
     }
     throw StateError('The factory "${factory.named}" is already present '
-        'in class "${id.id}"');
+        'in class "${id?.id == null ? '' : id!.id}"');
   }
 
   void addMethod(Method method) {
-    var existingMethod = _methods.firstWhere(
-        (element) =>
-            filled(element.id?.id) &&
-            filled(method.id?.id) &&
-            element.id.id == method.id.id &&
-            method.isSetter == element.isSetter &&
-            element.isGetter == element.isGetter,
-        orElse: () => null);
-    if (existingMethod == null) {
+    if (method.id == null) {
+      throw ArgumentError('Cannot add method with no id');
+    }
+    try {
+      _methods.firstWhere((element) =>
+          filled(element.id?.id) &&
+          filled(method.id?.id) &&
+          element.id!.id == method.id!.id &&
+          method.isSetter == element.isSetter &&
+          element.isGetter == element.isGetter);
+    } on StateError {
       method.parent = this;
       if (library != null) {
         method.library = library;
@@ -131,11 +148,12 @@ class DartClass extends NamedElement with DartType, Annotated {
     }
     throw StateError('The '
         '${method.isGetter ? 'getter' : method.isSetter ? 'setter' : 'method'} '
-        '"${method.id.id}" is already present in class "${id.id}"');
+        '"${method.id!.id}" is already present in class '
+        '"${id?.id == null ? '' : id!.id}"');
   }
 
   @override
-  void libraryUpdated(Library library) {
+  void libraryUpdated(Library? library) {
     super.libraryUpdated(library);
     for (var field in _fields) {
       field.library = library;
@@ -153,29 +171,41 @@ class DartClass extends NamedElement with DartType, Annotated {
 
   @override
   String generate() {
+    if (id?.id == null) {
+      throw ArgumentError('Cannot generate a class with no id');
+    }
     var ret = super.generate();
     if (ret != null) {
       return '${generateAnnotations()} $ret';
     }
     var buffer = StringBuffer();
-    buffer.writeln(generateAnnotations());
+    var annotation = generateAnnotations();
+    if (filled(annotation)) {
+      buffer.writeln(annotation);
+    }
     if (isAbstract) {
       buffer.writeKeyword(keywordAbstract);
     }
     buffer.writeKeyword(keywordClass);
-    buffer.writeIdentifier(id.id);
-    if (extend != null) {
+    buffer.writeIdentifier(id!.id);
+    if (extend != null && extend!.id != null) {
       buffer.writeKeyword(keywordExtends);
-      buffer.writeIdentifier(extend.id.id);
+      buffer.writeIdentifier(extend!.id!.id);
     }
     if (_withList.isNotEmpty) {
       buffer.writeKeyword(keywordWith);
-      final names = <String>[for (var def in _withList) def.id.id];
+      final names = <String>[
+        for (var def in _withList)
+          if (def.id != null) def.id!.id
+      ];
       buffer.writeNameList(names);
     }
     if (_implementList.isNotEmpty) {
       buffer.writeKeyword(keywordImplements);
-      final names = <String>[for (var def in _implementList) def.id.id];
+      final names = <String>[
+        for (var def in _implementList)
+          if (def.id != null) def.id!.id
+      ];
       buffer.writeNameList(names);
     }
     buffer.openBlock();
@@ -201,12 +231,15 @@ class DartClass extends NamedElement with DartType, Annotated {
       buffer.writeln('');
       var getterName = '';
       for (var method in _methods) {
-        if (!method.isSetter || method.id.id != getterName) {
+        if (method.id == null) {
+          continue;
+        }
+        if (!method.isSetter || method.id?.id != getterName) {
           buffer.writeln('');
         }
         if (method.isGetter) {
           buffer.write(method.generate());
-          getterName = method.id.id;
+          getterName = method.id!.id;
         } else {
           buffer.writeln(method.generate());
           getterName = '';
