@@ -3,10 +3,10 @@ import 'utils/keywords.dart';
 DartType typeDynamic = DartType()..type = keywordDynamic;
 
 class DartType {
-  String type;
+  String? type;
 
   @override
-  bool operator ==(other) => type == other.type;
+  bool operator ==(other) => other is DartType && type == other.type;
 
   @override
   int get hashCode => type.hashCode;
